% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_agg.R
\name{add_nodelen}
\alias{add_nodelen}
\title{Add node length and cumulative length (distance) downstream to node data}
\usage{
add_nodelen(nodedata, force = FALSE)
}
\arguments{
\item{nodedata}{As returned by \code{rt_read()}}

\item{force}{Logical, force via estimate if gaps exist?}
}
\description{
Add node length and cumulative length (distance) downstream to node data
}
