% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_agg.R
\name{add_offset}
\alias{add_offset}
\title{Add loc_offset to node data}
\usage{
add_offset(nodedata, reachdata)
}
\arguments{
\item{nodedata, reachdata}{As returned by \code{rt_read()}}
}
\description{
Add loc_offset to node data
}
