% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addVars}
\alias{addVars}
\title{Add variables to a data frame with given names}
\usage{
addVars(df, names, verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{names}{names of variables to add}

\item{verbose}{control messaging}
}
\description{
Helper for conflictsToCharacters function
}
