% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_agg.R
\name{all_sequential}
\alias{all_sequential}
\title{Check whether all node IDs are sequential --i.e. none are missing.}
\usage{
all_sequential(node_ids)
}
\arguments{
\item{node_ids}{Vector of IDs that should be sequential and complete.}
}
\description{
Very simple for now, but will be updated in future as node_id format is updated.
}
