% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{atts_df_fun}
\alias{atts_df_fun}
\title{Helper function to coerce attribute list to tabular}
\usage{
atts_df_fun(attlist)
}
\arguments{
\item{attlist}{as returnd by \code{ncdf4::ncatt_get()}}
}
\description{
Current approach is to paste everything together--in future
I may use list-valued columns.
}
