% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{attslist_to_dataframe}
\alias{attslist_to_dataframe}
\title{Convert pixc attributes to data frame.}
\usage{
attslist_to_dataframe(attslist)
}
\arguments{
\item{attslist}{list of attributes}
}
\description{
Convert pixc attributes to data frame.
}
