% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{badnodes}
\alias{badnodes}
\title{Returns a vector of worst-performing nodes (by error)}
\usage{
badnodes(valdata, variable = "width", n = 4, which = c("abs", "min",
  "max"))
}
\description{
Returns a vector of worst-performing nodes (by error)
}
