% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_rows2}
\alias{bind_rows2}
\title{A more flexible (but slower) version of dplyr::bind_rows}
\usage{
bind_rows2(dfList, addMissing = TRUE, verbose = FALSE)
}
\arguments{
\item{dfList}{a list of data.frames}

\item{addMissing}{Add columns if missing from a subset of dfList?}

\item{verbose}{Include messages about conversions?}
}
\description{
A more flexible (but slower) version of dplyr::bind_rows
}
