% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{check_ineff}
\alias{check_ineff}
\title{Check for inefficiencies in ncdf subset read function.}
\usage{
check_ineff(inds)
}
\arguments{
\item{inds}{Matrix giving rows and columns of indices}
}
\description{
Result should be close to 1.
}
