% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{flag_nodes}
\alias{flag_nodes}
\title{Flag nodes with questionable / unstable gdem truth}
\usage{
flag_nodes(dir, gdem1 = "rt_gdem.nc", gdem2 = "rt_gdem_dil2.nc",
  thresh = 0.15)
}
\arguments{
\item{dir}{Directory containing gdem rivertile netcdfs}

\item{gdem1, gdem2}{gdem rivertile netcdfs to compare}

\item{thresh}{Threshold for relative difference in widths, defaults to 0.15.}
}
\description{
Based on differences in "dilated" and non-dilated gdem.
}
