% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{getTileCorners}
\alias{getTileCorners}
\title{Get corners of tile polygon}
\usage{
getTileCorners(nadir1, nadir2, heading, xtstart = 4000, xtend = 64000,
  half = c("L", "R"))
}
\arguments{
\item{nadir1, nadir2}{length-2 vector giving longitude and latitude}

\item{heading}{satellite heading in degrees}

\item{xtstart, xtend}{Distance from nadir to start/end the swath, meters}

\item{half}{Which half of the tile: "L" for left, "R" for right}
}
\description{
Get corners of tile polygon
}
