% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-priordb.R
\name{getTilePolygons}
\alias{getTilePolygons}
\title{Get tiles as spatial frames POLYGON sfc}
\usage{
getTilePolygons(nadir1, nadir2, heading, half, xtstart = 4000,
  xtend = 64000)
}
\arguments{
\item{nadir1, nadir2}{Either a length-2 vector or a 2-column matrix giving lon, lat}

\item{heading}{satellite heading in degrees}

\item{half}{Which half of the tile: "L" for left, "R" for right}

\item{xtstart}{Distance from nadir to start/end the swath, meters}

\item{xtend}{Distance from nadir to start/end the swath, meters}
}
\description{
Returns a sfc with tile polygons
}
