% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{grab_nodes}
\alias{grab_nodes}
\alias{grab_nodebox}
\title{Filter a pixel cloud}
\usage{
grab_nodes(pixdf, nodeids, pcvdf = pixdf)

grab_nodebox(pixdf, nodeids, pcvdf = pixdf, dilate = 0)
}
\arguments{
\item{pixdf}{as returned by \code{pixc_read()} or \code{pixc_join()}}

\item{nodeids}{values of node_index to keep}

\item{pcvdf}{a data.frame that includes node and range/azimuth information,
e.g. as returned by \code{pixcvec_read()}. May be omitted if this information
is already in \code{pixdf}}

\item{dilate}{Number of pixels to extend beyond the supplied nodes
(default is 0).}
}
\description{
Only include pixels assigned to a given vector of nodes.
}
\section{Functions}{
\itemize{
\item \code{grab_nodebox}: Filter to a range/azimuth bounding box.
}}

