% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{join_pixc}
\alias{join_pixc}
\title{Join pixcvec to pixel cloud}
\usage{
join_pixc(dir, pcvname = "pcv.nc", pixcname = "pixel_cloud.nc")
}
\arguments{
\item{dir}{Directory containing pixel_cloud netcdfs}

\item{pcvname, pixcname}{Names of pixcvec and pixel cloud netcdf files}
}
\description{
Join pixcvec to pixel cloud
}
