% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{ncvar_ss}
\alias{ncvar_ss}
\title{Read in only a subset of a netcdf variable}
\usage{
ncvar_ss(nc, varid = NA, inds)
}
\arguments{
\item{nc}{netcdf with variable of interest}

\item{varid}{passed to \code{ncvar_get}}

\item{inds}{Vector or matrix giving indices for which the variable value is desired.}
}
\description{
Read in only a subset of a netcdf variable
}
