% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{orbit_read}
\alias{orbit_read}
\title{Read orbit locations}
\usage{
orbit_read(ncfile, as_sf = TRUE, maxpoints = 1000)
}
\arguments{
\item{ncfile}{netcdf file containing prior info}

\item{as_sf}{Convert to spatial frame object?}

\item{maxpoints}{Maximum number of points to use for line resolution.}
}
\description{
Returns a sf object with multilinestring geometry.
}
