% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{pixc_join}
\alias{pixc_join}
\title{Join pixcvec to pixel cloud}
\usage{
pixc_join(pixcdf, pcvdf, type = c("inner", "outer"))
}
\arguments{
\item{pixcdf}{as returned by \code{pixc_read()}}

\item{pcvdf}{as returned by \code{pixcvec_read()}}

\item{type}{What kind of join to perform? Default is "inner"}
}
\description{
Join pixcvec to pixel cloud
}
