% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{pixc_map}
\alias{pixc_map}
\alias{pixc_slantmap}
\title{Display a pixel cloud as a map.}
\usage{
pixc_map(pixdf, colorby = "classification", geoloc = c("best", "orig",
  "improved"), real_area = FALSE, water_frac = FALSE, plot = TRUE,
  maxpoints = 2500, ...)

pixc_slantmap(pixdf, colorby = "classification", maxpoints = 5000, ...)
}
\arguments{
\item{pixdf}{As returned by \code{pixc_read()}, possibly joined by
\code{pixc_join()}}

\item{colorby}{variable name in \code{pixcdf} to use for coloring}

\item{geoloc}{Which geolocation to use: "best" (default) chooses based on
which columns are available in \code{pixcdf}.}

\item{real_area}{If \code{TRUE}, use display actual pixel areas. Note that
this will take considerably more computation time.}

\item{water_frac}{Scale pixel size by \code{water_frac} column?}

\item{plot}{if FALSE return the plot data instead of generating the plot.}

\item{maxpoints}{Maximum number of points to allow. Meant to encourage
filtering so as not to overwhelm the renderer.}

\item{...}{passed to \code{ggplot2::geom_point} or \code{ggforce::geom_circle}}
}
\description{
Display a pixel cloud as a map.

Slant-plane map of pixel cloud
}
\section{Functions}{
\itemize{
\item \code{pixc_slantmap}: Slant-plane map of pixel cloud
}}

