% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{pixc_read}
\alias{pixc_read}
\title{Read data from a pixel_cloud netcdf file}
\usage{
pixc_read(ncfile, group = c("pixel_cloud", "tvp", "noise"),
  latlim = c(-90, 90), lonlim = c(-180, 180), keep_na_vars = FALSE)
}
\arguments{
\item{ncfile}{a pixel_cloud netcdf file}

\item{group}{Which group to get data from: "pixel_cloud", "tvp", or "noise"}

\item{latlim}{}
}
\description{
Read data from a pixel_cloud netcdf file
}
