% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{pixcvec_read}
\alias{pixcvec_read}
\title{Read a pixcvec from a netcdf file}
\usage{
pixcvec_read(ncfile, keep_na_vars = FALSE)
}
\arguments{
\item{ncfile}{PIXCVEC netcdf file}

\item{keep_na_vars}{Keep variables that only contain missing values?}
}
\description{
Read a pixcvec from a netcdf file
}
