% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-priordb.R
\name{priorcl_read}
\alias{priorcl_read}
\title{Read prior database for centerline info}
\usage{
priorcl_read(ncfile, nodeids = NULL, reachids = NULL, as_sf = TRUE)
}
\arguments{
\item{ncfile}{netcdf file containing prior info}

\item{nodeids}{optional vector of node indices}

\item{reachids}{optional vector of reach indices}

\item{as_sf}{Convert to spatial frame object?}
}
\description{
Returns a sf object with linestring geometry.
}
