% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp-test.R
\name{rt_hyptest}
\alias{rt_hyptest}
\title{Test uncertainty estimates against chi-square distribution}
\usage{
rt_hyptest(valdata, debias = TRUE)
}
\arguments{
\item{valdata}{A data.frame as returned by \code{rt_valdata()}}

\item{debias}{Remove effect of bias? Defaults to TRUE}
}
\description{
Test uncertainty estimates against chi-square distribution
}
