% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rt_val_nodeseries}
\alias{rt_val_nodeseries}
\title{Plot predictions and uncertainty bounds across nodes}
\usage{
rt_val_nodeseries(dir, variable = "height", err_only = TRUE, ...)
}
\arguments{
\item{dir}{directory containing riverobs output including gdem truth}

\item{variable}{Which variable to plot?}

\item{err_only}{Only plot the errors (subtract off truth value)?}

\item{...}{Passed to \code{rt_valdata()}}
}
\description{
Plot predictions and uncertainty bounds across nodes
}
