% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{rt_valdata_df}
\alias{rt_valdata_df}
\title{Create validation data from two node-level rivertile data.frames.}
\usage{
rt_valdata_df(obs, truth, time_round_digits = -2)
}
\description{
Create validation data from two node-level rivertile data.frames.
}
