% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitPiece}
\alias{splitPiece}
\title{Copied from markstats package}
\usage{
splitPiece(strvec, split, piece, ...)
}
\arguments{
\item{strvec}{Character vector}

\item{split}{Character to split by}

\item{piece}{of split element to keep (numeric)}

\item{...}{Passed to \code{strsplit()}}
}
\description{
Copied from markstats package
}
