% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{to_latlon}
\alias{to_latlon}
\title{Convert linear distance to lat/lon distance}
\usage{
to_latlon(x, lat)
}
\arguments{
\item{x}{distance in meters}

\item{lat}{latitude, used for determining degree length}
}
\description{
Helper for map_node_pixc.
}
