% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{use_condaenv}
\alias{use_condaenv}
\title{Fix for reticulate::use_condaenv}
\usage{
use_condaenv(condaenv, conda = "auto", required = FALSE)
}
\description{
Previously ran into a warning about condition having length > 1
}
