% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{val_map_node}
\alias{val_map_node}
\title{Map a given number of nodes' pixcvec locations--gdem versus rivertile}
\usage{
val_map_node(dir, nodes = badnodes(rt_valdata(dir)), pcv1 = "pcv.nc",
  pcv2 = "pcv_gdem.nc", maxpixels = 1000, leaflet = TRUE, ...)
}
\arguments{
\item{nodes}{A vector of node_id's, defaults to worst nodes using \code{badnodes()}}

\item{pcv1}{Name of first pixcvec netcdf}

\item{pcv2}{Name of second pixcvec netcdf}

\item{maxpixels}{Maximum number of pixels to plot, as a random sample if necessary.}

\item{leaflet}{Use interactive leaflet plot? Otherwise uses ggplot.}

\item{...}{Passed to \code{rt_valdata}}

\item{valdata}{data.frame as returned by \code{rt_valdata()}}
}
\description{
Map a given number of nodes' pixcvec locations--gdem versus rivertile
}
