% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_agg.R
\name{add_nodelen}
\alias{add_nodelen}
\title{Add node length and cumulative length (distance) downstream to node data}
\usage{
add_nodelen(nodedata, reachdata)
}
\arguments{
\item{nodedata, reachdata}{As returned by \code{rt_read()}}
}
\description{
Add node length and cumulative length (distance) downstream to node data
}
