% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{priorreach_read}
\alias{priorreach_read}
\title{Read prior database for node info}
\usage{
priorreach_read(ncfile, reachids = NULL)
}
\arguments{
\item{ncfile}{netcdf file containing prior info}

\item{nodeids}{vector of node indices}
}
\description{
Unclear when this function would be useful.
}
