% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_agg.R
\name{reach_height_lm}
\alias{reach_height_lm}
\title{Weighted linear model for height as a function of distance downstream}
\usage{
reach_height_lm(node_h, node_h_u, node_x, loc_offset, weight = TRUE)
}
\arguments{
\item{node_h}{Vector of node heights}

\item{node_h_u}{Vector of node height uncertainty (1-sigma)}

\item{node_x}{Vector of node downstream distance}

\item{loc_offset}{From reach-level rivertile, distance from observed to prior
reach center}

\item{weight}{Use weighted least-squares? Default is TRUE.}
}
\description{
Used for computing reach-level heights and slopes with corresponding uncertainty.
}
