% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-products.R
\name{rt_read}
\alias{rt_read}
\title{Get data from a rivertile netcdf}
\usage{
rt_read(ncfile, group = c("nodes", "reaches"), keep_na_vars = FALSE,
  as_sf = FALSE)
}
\arguments{
\item{ncfile}{A rivertile netcdf file}

\item{group}{Which group to get data from: "nodes" or "reaches}

\item{keep_na_vars}{Keep variables that only contain missing values?}

\item{as_sf}{Convert to a spatial frame?}
}
\description{
Get data from a rivertile netcdf
}
